Use 12:00 as a reference point for simple dates instead of midnight. (#4025)
In locales where DST timezone changes occur on midnight (like Brazil) the calendar becomes confused which date to show. On DST ends, the date sprung back from 2010-10-17 00:00 to 2010-10-16 23:00 in Brazil. Thus, the date 2010-10-16 was shown two times. Using 12 as the default hour for dates does circumvent this issue. If DST changes occur around noon or the time picker functionality is used, this still breaks however.